Update structure to enable usage as subapplication.
Move fubar app to main directory in order to have fubar.app.src in src
directory directly in the root folder so that fubar may be used as
a dependency under rebar.

The file rebar.config has been updated to accomodate this change. The rel
folder has been removed and releases should be handled in the application
that uses fubar.